pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt-knowledge-builder"
dynamic = ["version"]
description = "Compile mixed documents into upload-ready Custom GPT Knowledge packages."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "GPT Knowledge Builder" },
]
keywords = [
  "custom-gpt",
  "knowledge-base",
  "document-processing",
  "windows-app",
  "tkinter",
  "openai",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Win32 (MS Windows)",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Text Processing",
]
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
scan-docs = "knowledge_builder.scan_docs:main"

[project.gui-scripts]
gpt-knowledge-builder = "knowledge_builder.gui_entry:main"

[project.optional-dependencies]
ai = [
  "openai>=1.68.0",
]
extractors = [
  "PyMuPDF>=1.24.0",
  "pdfplumber>=0.11.0",
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
  "beautifulsoup4>=4.12.0",
  "openpyxl>=3.1.0",
  "python-pptx>=1.0.0",
]
ocr = [
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
]
dev = [
  "pytest>=8.0.0",
]
windows-build = [
  "pyinstaller>=6.11.0",
]

[project.urls]
Homepage = "https://github.com/AboveWireless/gpt-knowledge-builder"
Repository = "https://github.com/AboveWireless/gpt-knowledge-builder"
Issues = "https://github.com/AboveWireless/gpt-knowledge-builder/issues"
Changelog = "https://github.com/AboveWireless/gpt-knowledge-builder/blob/main/CHANGELOG.md"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
include = ["knowledge_builder*"]
exclude = ["tests*", "packaging*"]

[tool.setuptools.dynamic]
version = { attr = "knowledge_builder.version.APP_VERSION" }
